GHC.Platform

Description

A description of the platform we're compiling for.

Documentation

data PlatformMini

Contains the bare-bones arch and os information. This isn't enough for code gen, but useful for tasks where we can fall back upon the host platform, as this is all we know about the host platform.

Constructors

PlatformMini

Fields

platformMini_arch :: Arch

platformMini_os :: OS

data PlatformWordSize

Constructors

PW4
A 32-bit platform

PW8
A 64-bit platform

data Platform

Contains enough information for the native code generator to emit code for this platform.

Constructors

Platform

Fields

platformMini :: PlatformMini

platformWordSize :: PlatformWordSize

platformUnregisterised :: Bool

platformHasGnuNonexecStack :: Bool

platformHasIdentDirective :: Bool

platformHasSubsectionsViaSymbols :: Bool

platformIsCrossCompiling :: Bool

platformArch :: Platform -> Arch

Legacy accessor

platformOS :: Platform -> OS

Legacy accessor

data Arch

Architectures that the native code generator knows about. TODO: It might be nice to extend these constructors with information about what instruction set extensions an architecture might support.

Constructors

ArchUnknown

ArchX86

ArchX86_64

ArchPPC

ArchPPC_64

Fields

ppc_64ABI :: PPC_64ABI

ArchS390X

ArchSPARC

ArchSPARC64

ArchARM

Fields

armISA :: ArmISA

armISAExt :: [ArmISAExt]

armABI :: ArmABI

ArchAArch64

ArchAlpha

ArchMipseb

ArchMipsel

ArchJavaScript

data OS

Operating systems that the native code generator knows about. Having OSUnknown should produce a sensible default, but no promises.

Constructors

OSUnknown

OSLinux

OSDarwin

OSSolaris2

OSMinGW32

OSFreeBSD

OSDragonFly

OSOpenBSD

OSNetBSD

OSKFreeBSD

OSHaiku

OSQNXNTO

OSAIX

OSHurd

data ArmISA

ARM Instruction Set Architecture, Extensions and ABI

Constructors

ARMv5

ARMv6

ARMv7

data ArmISAExt

Constructors

VFPv2

VFPv3

VFPv3D16

NEON

IWMMX2

data ArmABI

Constructors

SOFT

SOFTFP

HARD

data PPC_64ABI

PowerPC 64-bit ABI

Constructors

ELF_V1

ELF_V2

target32Bit :: Platform -> Bool

This predicate tells us whether the platform is 32-bit.

isARM :: Arch -> Bool

osElfTarget :: OS -> Bool

This predicate tells us whether the OS supports ELF-like shared libraries.

osMachOTarget :: OS -> Bool

This predicate tells us whether the OS support Mach-O shared libraries.

osSubsectionsViaSymbols :: OS -> Bool

platformUsesFrameworks :: Platform -> Bool

platformWordSizeInBytes :: Platform -> Int

platformWordSizeInBits :: Platform -> Int

data PlatformMisc

Platform-specific settings formerly hard-coded in Config.hs.

These should probably be all be triaged whether they can be computed from other settings or belong in another another place (like Platform above).

Constructors

PlatformMisc

Fields

platformMisc_targetPlatformString :: String

platformMisc_integerLibrary :: String

platformMisc_integerLibraryType :: IntegerLibrary

platformMisc_ghcWithInterpreter :: Bool

platformMisc_ghcWithNativeCodeGen :: Bool

platformMisc_ghcWithSMP :: Bool

platformMisc_ghcRTSWays :: String

platformMisc_tablesNextToCode :: Bool

Determines whether we will be compiling info tables that reside just before the entry code, or with an indirection to the entry code. See TABLES_NEXT_TO_CODE in includesrtsstorage/InfoTables.h.

platformMisc_leadingUnderscore :: Bool

platformMisc_libFFI :: Bool

platformMisc_ghcThreaded :: Bool

platformMisc_ghcDebugged :: Bool

platformMisc_ghcRtsWithLibdw :: Bool

platformMisc_llvmTarget :: String

data IntegerLibrary

Constructors

IntegerGMP

IntegerSimple

stringEncodeArch :: Arch -> String

See Note [Platform Syntax].

stringEncodeOS :: OS -> String

See Note [Platform Syntax].