GHC.Platform.ArchOS

Description

Platform architecture and OS

We need it in ghc-boot because ghc-pkg needs it.

Documentation

data ArchOS

Platform architecture and OS.

Constructors

ArchOS

Fields

archOS_arch :: Arch

archOS_OS :: OS

data Arch

Architectures

TODO: It might be nice to extend these constructors with information about what instruction set extensions an architecture might support.

Constructors

ArchUnknown

ArchX86

ArchX86_64

ArchPPC

ArchPPC_64 PPC_64ABI

ArchS390X

ArchSPARC

ArchSPARC64

ArchARM ArmISA [ArmISAExt] ArmABI

ArchAArch64

ArchAlpha

ArchMipseb

ArchMipsel

ArchRISCV64

ArchJavaScript

data OS

Operating systems.

Using OSUnknown to generate code should produce a sensible default, but no promises.

Constructors

OSUnknown

OSLinux

OSDarwin

OSSolaris2

OSMinGW32

OSFreeBSD

OSDragonFly

OSOpenBSD

OSNetBSD

OSKFreeBSD

OSHaiku

OSQNXNTO

OSAIX

OSHurd

data ArmISA

ARM Instruction Set Architecture

Constructors

ARMv5

ARMv6

ARMv7

data ArmISAExt

ARM extensions

Constructors

VFPv2

VFPv3

VFPv3D16

NEON

IWMMX2

data ArmABI

ARM ABI

Constructors

SOFT

SOFTFP

HARD

data PPC_64ABI

PowerPC 64-bit ABI

Constructors

ELF_V1
PowerPC64

ELF_V2
PowerPC64 LE

stringEncodeArch :: Arch -> String

See Note [Platform Syntax].

stringEncodeOS :: OS -> String

See Note [Platform Syntax].