GHC.Platform

Description

A description of the platform we're compiling for.

Documentation

data PlatformMini

Contains the bare-bones arch and os information. This isn't enough for code gen, but useful for tasks where we can fall back upon the host platform, as this is all we know about the host platform.

Constructors

PlatformMini

Fields

platformMini_arch :: Arch

platformMini_os :: OS

data PlatformWordSize

Constructors

PW4
A 32-bit platform

PW8
A 64-bit platform

data Platform

Contains enough information for the native code generator to emit code for this platform.

Constructors

Platform

Fields

platformMini :: !PlatformMini

platformWordSize :: !PlatformWordSize

Word size

platformByteOrder :: !ByteOrder

Byte order (endianness)

platformUnregisterised :: !Bool

platformHasGnuNonexecStack :: !Bool

platformHasIdentDirective :: !Bool

platformHasSubsectionsViaSymbols :: !Bool

platformIsCrossCompiling :: !Bool

platformLeadingUnderscore :: !Bool

Symbols need underscore prefix

platformTablesNextToCode :: !Bool

Determines whether we will be compiling info tables that reside just before the entry code, or with an indirection to the entry code. See TABLES_NEXT_TO_CODE in includesrtsstorage/InfoTables.h.

platformArch :: Platform -> Arch

Legacy accessor

platformOS :: Platform -> OS

Legacy accessor

data Arch

Architectures that the native code generator knows about. TODO: It might be nice to extend these constructors with information about what instruction set extensions an architecture might support.

Constructors

ArchUnknown

ArchX86

ArchX86_64

ArchPPC

ArchPPC_64

Fields

ppc_64ABI :: PPC_64ABI

ArchS390X

ArchSPARC

ArchSPARC64

ArchARM

Fields

armISA :: ArmISA

armISAExt :: [ArmISAExt]

armABI :: ArmABI

ArchAArch64

ArchAlpha

ArchMipseb

ArchMipsel

ArchJavaScript

data OS

Operating systems that the native code generator knows about. Having OSUnknown should produce a sensible default, but no promises.

Constructors

OSUnknown

OSLinux

OSDarwin

OSSolaris2

OSMinGW32

OSFreeBSD

OSDragonFly

OSOpenBSD

OSNetBSD

OSKFreeBSD

OSHaiku

OSQNXNTO

OSAIX

OSHurd

data ArmISA

ARM Instruction Set Architecture, Extensions and ABI

Constructors

ARMv5

ARMv6

ARMv7

data ArmISAExt

Constructors

VFPv2

VFPv3

VFPv3D16

NEON

IWMMX2

data ArmABI

Constructors

SOFT

SOFTFP

HARD

data PPC_64ABI

PowerPC 64-bit ABI

Constructors

ELF_V1

ELF_V2

data ByteOrder

Byte ordering.

Constructors

BigEndian
most-significant-byte occurs in lowest address.

LittleEndian
least-significant-byte occurs in lowest address.

target32Bit :: Platform -> Bool

This predicate tells us whether the platform is 32-bit.

isARM :: Arch -> Bool

osElfTarget :: OS -> Bool

This predicate tells us whether the OS supports ELF-like shared libraries.

osMachOTarget :: OS -> Bool

This predicate tells us whether the OS support Mach-O shared libraries.

osSubsectionsViaSymbols :: OS -> Bool

platformUsesFrameworks :: Platform -> Bool

platformWordSizeInBytes :: Platform -> Int

platformWordSizeInBits :: Platform -> Int

platformMinInt :: Platform -> Integer

Minimum representable Int value for the given platform

platformMaxInt :: Platform -> Integer

Maximum representable Int value for the given platform

platformMaxWord :: Platform -> Integer

Maximum representable Word value for the given platform

platformInIntRange :: Platform -> Integer -> Bool

Test if the given Integer is representable with a platform Int

platformInWordRange :: Platform -> Integer -> Bool

Test if the given Integer is representable with a platform Word

platformCConvNeedsExtension :: Platform -> Bool

For some architectures the C calling convention is that any integer shorter than 64 bits is replaced by its 64 bits representation using sign or zero extension.

data PlatformMisc

Platform-specific settings formerly hard-coded in Config.hs.

These should probably be all be triaged whether they can be computed from other settings or belong in another another place (like Platform above).

Constructors

PlatformMisc

Fields

platformMisc_targetPlatformString :: String

platformMisc_ghcWithInterpreter :: Bool

platformMisc_ghcWithSMP :: Bool

platformMisc_ghcRTSWays :: String

platformMisc_libFFI :: Bool

platformMisc_ghcThreaded :: Bool

platformMisc_ghcDebugged :: Bool

platformMisc_ghcRtsWithLibdw :: Bool

platformMisc_llvmTarget :: String

stringEncodeArch :: Arch -> String

See Note [Platform Syntax].

stringEncodeOS :: OS -> String

See Note [Platform Syntax].

data SseVersion

x86 SSE instructions

Constructors

SSE1

SSE2

SSE3

SSE4

SSE42

data BmiVersion

x86 BMI (bit manipulation) instructions

Constructors

BMI1

BMI2